% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\docType{data}
\name{URL}
\alias{URL}
\title{List containing data world urls}
\format{
An object of class \code{list} of length 3.
}
\usage{
URL
}
\description{
List containing data world urls
}
\keyword{datasets}
