% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{getData}
\alias{getData}
\title{Load data and its metadata from data world.}
\usage{
getData(dataset_name)
}
\arguments{
\item{dataset_name}{Any of the following shorthand names:
\itemize{
\item "hhs" - Household surveys. hh_survey_combined from https://data.world/rare/hh-surveys
\item "fish" - Fish surveys. fish_surveys_all from https://data.world/rare/fish-surveys
\item "benthic" - Benthic surveys. BenthicMaster from https://data.world/rare/benthic-surveys
\item "ourfish" - OurFish catch data. join_ourfish_footprint_fishbase from https://data.world/rare/fisheries-dashboard
}}
}
\value{
A list containing the requested data, its column metadata, and source url.
}
\description{
Load data and its metadata from data world.
}
